% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifierEvaluation.R
\name{BinaryClassifierResult}
\alias{BinaryClassifierResult}
\title{ClassifierResult class}
\description{
ClassifierResult class

ClassifierResult class
}
\details{
result
}
\keyword{distributions}
\section{Super class}{
\code{\link[ClassifierResult:ClassifierResult]{ClassifierResult::ClassifierResult}} -> \code{BinaryClassifierResult}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BinaryClassifierResult$new()}}
\item \href{#method-clone}{\code{BinaryClassifierResult$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ClassifierResult" data-topic="ClassifierResult" data-id="getInfoStats">}\href{../../ClassifierResult/html/ClassifierResult.html#method-getInfoStats}{\code{ClassifierResult::ClassifierResult$getInfoStats()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ClassifierResult" data-topic="ClassifierResult" data-id="plotRoc">}\href{../../ClassifierResult/html/ClassifierResult.html#method-plotRoc}{\code{ClassifierResult::ClassifierResult$plotRoc()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Creates a distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinaryClassifierResult$new(data, classes)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{predictionProbabilities}}{a matrix of probabilities with rows for every sample, and colums for every possible class}

\item{\code{actual}}{the gold standard output as a vector of outcomes for each sample}

\item{\code{...}}{passed to pdfFunction and centileFunction}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinaryClassifierResult$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
