% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassifierComparison.R
\name{ClassifierComparison}
\alias{ClassifierComparison}
\title{ClassifierComparison class}
\description{
result
}
\keyword{distributions}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{- the classifier prediction probabilities and outcomes
adds a classifier result to the list}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{data}}{- the classifier prediction probabilities and outcomes
adds a classifier result to the list}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ClassifierComparison$new()}}
\item \href{#method-withClassifier}{\code{ClassifierComparison$withClassifier()}}
\item \href{#method-withResult}{\code{ClassifierComparison$withResult()}}
\item \href{#method-compareMaxValues}{\code{ClassifierComparison$compareMaxValues()}}
\item \href{#method-compareDistributionStats}{\code{ClassifierComparison$compareDistributionStats()}}
\item \href{#method-getInfoStats}{\code{ClassifierComparison$getInfoStats()}}
\item \href{#method-plotRocs}{\code{ClassifierComparison$plotRocs()}}
\item \href{#method-clone}{\code{ClassifierComparison$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassifierComparison$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-withClassifier"></a>}}
\subsection{Method \code{withClassifier()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassifierComparison$withClassifier(
  parameterList,
  predicitionProbabilities,
  obs
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parameterList}}{a list of parameters used to generate this classifier result}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the ClassifierComparison
adds a classifier result to the list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-withResult"></a>}}
\subsection{Method \code{withResult()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassifierComparison$withResult(parameterList, classifierResult)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parameterList}}{a list of parameters used to generate this classifier result}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the ClassifierComparison
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compareMaxValues"></a>}}
\subsection{Method \code{compareMaxValues()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassifierComparison$compareMaxValues(statisticVars, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compareDistributionStats"></a>}}
\subsection{Method \code{compareDistributionStats()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassifierComparison$compareDistributionStats(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getInfoStats"></a>}}
\subsection{Method \code{getInfoStats()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassifierComparison$getInfoStats()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotRocs"></a>}}
\subsection{Method \code{plotRocs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassifierComparison$plotRocs(
  titleFromParamsFn = function(params) {     paste(names(params), "=", params, collapse
    = "; ") }
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassifierComparison$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
