% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MulticlassClassifierResult.R
\name{MulticlassClassifierResult}
\alias{MulticlassClassifierResult}
\title{ClassifierResult class}
\description{
ClassifierResult class

ClassifierResult class
}
\details{
result
}
\keyword{distributions}
\section{Super class}{
\code{\link[ClassifierResult:ClassifierResult]{ClassifierResult::ClassifierResult}} -> \code{MulticlassClassifierResult}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MulticlassClassifierResult$new()}}
\item \href{#method-getBinaryClassifiers}{\code{MulticlassClassifierResult$getBinaryClassifiers()}}
\item \href{#method-getClasses}{\code{MulticlassClassifierResult$getClasses()}}
\item \href{#method-plotRoc}{\code{MulticlassClassifierResult$plotRoc()}}
\item \href{#method-clone}{\code{MulticlassClassifierResult$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ClassifierResult" data-topic="ClassifierResult" data-id="distributionStats">}\href{../../ClassifierResult/html/ClassifierResult.html#method-distributionStats}{\code{ClassifierResult::ClassifierResult$distributionStats()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ClassifierResult" data-topic="ClassifierResult" data-id="getInfoStats">}\href{../../ClassifierResult/html/ClassifierResult.html#method-getInfoStats}{\code{ClassifierResult::ClassifierResult$getInfoStats()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ClassifierResult" data-topic="ClassifierResult" data-id="maxValue">}\href{../../ClassifierResult/html/ClassifierResult.html#method-maxValue}{\code{ClassifierResult::ClassifierResult$maxValue()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ClassifierResult" data-topic="ClassifierResult" data-id="maxValues">}\href{../../ClassifierResult/html/ClassifierResult.html#method-maxValues}{\code{ClassifierResult::ClassifierResult$maxValues()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Creates a distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MulticlassClassifierResult$new(data, classes)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{passed to pdfFunction and centileFunction}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getBinaryClassifiers"></a>}}
\subsection{Method \code{getBinaryClassifiers()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MulticlassClassifierResult$getBinaryClassifiers()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getClasses"></a>}}
\subsection{Method \code{getClasses()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MulticlassClassifierResult$getClasses()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotRoc"></a>}}
\subsection{Method \code{plotRoc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MulticlassClassifierResult$plotRoc()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MulticlassClassifierResult$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
