% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributionFunctions.R
\name{NormalDistribution}
\alias{NormalDistribution}
\title{Get a normal distribution wrapper}
\description{
Get a normal distribution wrapper

Get a normal distribution wrapper
}
\keyword{distributions}
\section{Super class}{
\code{\link[ClassifierResult:Distribution]{ClassifierResult::Distribution}} -> \code{NormalDistribution}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{mu}}{the mean of the normal distribuition}

\item{\code{sigma}}{the sigma value for the distribution}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{NormalDistribution$new()}}
\item \href{#method-label}{\code{NormalDistribution$label()}}
\item \href{#method-theoreticalEntropy}{\code{NormalDistribution$theoreticalEntropy()}}
\item \href{#method-clone}{\code{NormalDistribution$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ClassifierResult" data-topic="Distribution" data-id="bigP">}\href{../../ClassifierResult/html/Distribution.html#method-bigP}{\code{ClassifierResult::Distribution$bigP()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ClassifierResult" data-topic="Distribution" data-id="getInverseCdf">}\href{../../ClassifierResult/html/Distribution.html#method-getInverseCdf}{\code{ClassifierResult::Distribution$getInverseCdf()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ClassifierResult" data-topic="Distribution" data-id="getPdf">}\href{../../ClassifierResult/html/Distribution.html#method-getPdf}{\code{ClassifierResult::Distribution$getPdf()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ClassifierResult" data-topic="Distribution" data-id="inverse">}\href{../../ClassifierResult/html/Distribution.html#method-inverse}{\code{ClassifierResult::Distribution$inverse()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ClassifierResult" data-topic="Distribution" data-id="p">}\href{../../ClassifierResult/html/Distribution.html#method-p}{\code{ClassifierResult::Distribution$p()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ClassifierResult" data-topic="Distribution" data-id="plot">}\href{../../ClassifierResult/html/Distribution.html#method-plot}{\code{ClassifierResult::Distribution$plot()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ClassifierResult" data-topic="Distribution" data-id="q">}\href{../../ClassifierResult/html/Distribution.html#method-q}{\code{ClassifierResult::Distribution$q()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ClassifierResult" data-topic="Distribution" data-id="sample">}\href{../../ClassifierResult/html/Distribution.html#method-sample}{\code{ClassifierResult::Distribution$sample()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ClassifierResult" data-topic="Distribution" data-id="twoDp">}\href{../../ClassifierResult/html/Distribution.html#method-twoDp}{\code{ClassifierResult::Distribution$twoDp()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
get a new normal distribution parameterised by mean and sd
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NormalDistribution$new(mean = runif(1, -2, 2), sd = runif(1, 0.5, 5))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mean}}{the mean}

\item{\code{sd}}{the sd}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a NormalDistribution object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-label"></a>}}
\subsection{Method \code{label()}}{
gets a label for this distribution based on the parameters passed
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NormalDistribution$label()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-theoreticalEntropy"></a>}}
\subsection{Method \code{theoreticalEntropy()}}{
calculates the theoretical entropy
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NormalDistribution$theoreticalEntropy()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NormalDistribution$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
