% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinaryClassifierResult.R
\name{BinaryClassifierResult}
\alias{BinaryClassifierResult}
\title{ClassifierResult class}
\description{
ClassifierResult class

ClassifierResult class
}
\details{
result
}
\keyword{distributions}
\section{Super class}{
\code{\link[ClassifierResult:ClassifierResult]{ClassifierResult::ClassifierResult}} -> \code{BinaryClassifierResult}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{positive}}{- the "positive" class}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BinaryClassifierResult$new()}}
\item \href{#method-setPositive}{\code{BinaryClassifierResult$setPositive()}}
\item \href{#method-plotRoc}{\code{BinaryClassifierResult$plotRoc()}}
\item \href{#method-plotDensity}{\code{BinaryClassifierResult$plotDensity()}}
\item \href{#method-plotStats}{\code{BinaryClassifierResult$plotStats()}}
\item \href{#method-plotPR}{\code{BinaryClassifierResult$plotPR()}}
\item \href{#method-distributionStats}{\code{BinaryClassifierResult$distributionStats()}}
\item \href{#method-clone}{\code{BinaryClassifierResult$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ClassifierResult" data-topic="ClassifierResult" data-id="getInfoStats">}\href{../../ClassifierResult/html/ClassifierResult.html#method-getInfoStats}{\code{ClassifierResult::ClassifierResult$getInfoStats()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ClassifierResult" data-topic="ClassifierResult" data-id="maxValue">}\href{../../ClassifierResult/html/ClassifierResult.html#method-maxValue}{\code{ClassifierResult::ClassifierResult$maxValue()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ClassifierResult" data-topic="ClassifierResult" data-id="maxValues">}\href{../../ClassifierResult/html/ClassifierResult.html#method-maxValues}{\code{ClassifierResult::ClassifierResult$maxValues()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Creates a distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinaryClassifierResult$new(data, classes)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{predictionProbabilities}}{a matrix of probabilities with rows for every sample, and colums for every possible class}

\item{\code{actual}}{the gold standard output as a vector of outcomes for each sample}

\item{\code{...}}{passed to pdfFunction and centileFunction}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setPositive"></a>}}
\subsection{Method \code{setPositive()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinaryClassifierResult$setPositive(className)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotRoc"></a>}}
\subsection{Method \code{plotRoc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinaryClassifierResult$plotRoc()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotDensity"></a>}}
\subsection{Method \code{plotDensity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinaryClassifierResult$plotDensity()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotStats"></a>}}
\subsection{Method \code{plotStats()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinaryClassifierResult$plotStats(
  statisticVars = vars(neg_pred_value, pos_pred_value, specificity, sensitivity,
    precision, recall, accuracy, f1, mcc, informedness, I)
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotPR"></a>}}
\subsection{Method \code{plotPR()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinaryClassifierResult$plotPR()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-distributionStats"></a>}}
\subsection{Method \code{distributionStats()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinaryClassifierResult$distributionStats()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinaryClassifierResult$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
