% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifierEvaluation.R
\name{ClassifierResult}
\alias{ClassifierResult}
\title{ClassifierResult class}
\description{
ClassifierResult class

ClassifierResult class
}
\details{
result
}
\keyword{distributions}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{- the classifier prediction probabilities and outcomes}

\item{\code{classes}}{- the classes of the outcome as a vector}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ClassifierResult$new()}}
\item \href{#method-getInfoStats}{\code{ClassifierResult$getInfoStats()}}
\item \href{#method-plotRoc}{\code{ClassifierResult$plotRoc()}}
\item \href{#method-clone}{\code{ClassifierResult$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Creates a distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassifierResult$new(data, classes)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{predictionProbabilities}}{a matrix of probabilities with rows for every sample, and colums for every possible class}

\item{\code{actual}}{the gold standard output as a vector of outcomes for each sample}

\item{\code{...}}{passed to pdfFunction and centileFunction}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getInfoStats"></a>}}
\subsection{Method \code{getInfoStats()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassifierResult$getInfoStats()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotRoc"></a>}}
\subsection{Method \code{plotRoc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassifierResult$plotRoc()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassifierResult$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
